1) added default path to UserDirectory 2) Exceptions and Loggs are fixed

+++ src/EvidentInstruction/Models/TextFile.cs
@@ -5,11 +5,10 @@ using System.Threading;
 using EvidentInstruction.Exceptions;
 using EvidentInstruction.Helpers;
 
-
 namespace EvidentInstruction.Models
 {
     public class TextFile : IFile, IDisposable
-    {
+    { 
         private ThreadLocal<UserDirectory> _userDirectory { get; } = new ThreadLocal<UserDirectory>();
         public UserDirectory UserDirectory
         {
@@ -24,25 +23,27 @@ namespace EvidentInstruction.Models
         {
             UserDirectory = new UserDirectory();
         }
-        private bool CheckFileExtension(string filename)
+        protected bool CheckFileExtension(string filename)
         {
             string extension = Path.GetExtension(filename);
-            string extMustBeTxt = FileExtensions.txt;
+            string extMustBeTxt = FileExtensions.TXT;
             bool result = (extension == extMustBeTxt) ? true : false;
             return result;
         }
         public bool IsExist(string filename, string path)
         {
+            if (string.IsNullOrWhiteSpace(path)) path = UserDirectory.Get();
             string fullpath = Path.Combine(path, filename);
             bool result = (System.IO.File.Exists(fullpath)) ? true : false;
             return result;
         }
         public void DownloadFileByURL(string url, string filename, string pathToSave)
         {
+            if (string.IsNullOrWhiteSpace(pathToSave)) pathToSave = UserDirectory.Get();
             if (string.IsNullOrWhiteSpace(filename))
             {
-                Log.Logger.Warning("Отсутствует имя файла");
-                throw new ArgumentException("Имя файла " + filename + " отсутствует");
+                Log.Logger.Warning("Имя файла отсутствует");
+                throw new ArgumentException("Имя файла отсутствует");
             }
             else
             {
@@ -58,30 +59,30 @@ namespace EvidentInstruction.Models
                             bool isExist = IsExist(filename, pathToSave);
                             if (isExist)
                             {
-                                Log.Logger.Information("Файл " + filename + " скачан ");
+                                Log.Logger.Information("Файл  \"{0}\" был скачан в директорию \"{1}\" ", filename, pathToSave);
                             }
                             else
                             {
-                                Log.Logger.Warning("Файл " + filename + " не скачан");
-                                throw new FileNotFoundException("Файл " + filename + " скачан");
+                                Log.Logger.Warning("Файл \"{0}\" не скачан",filename);
+                                throw new FileNotFoundException("Файл \"{0}\" не скачан", filename);
                             }
                         }
                     }
                     catch (WebException e)
                     {
-                        Log.Logger.Warning(e.Message);
+                        Log.Logger.Warning("Файл \"{0}\" не скачан по ошибке {1}", filename,e.Message);
                     }
                 }
                 else
                 {
-                    throw new ValidFileNameException("Проверьте, что файл " + filename + " имеет расширение .txt");
-                    Log.Logger.Warning("Проверьте, что файл " + filename + " имеет расширение .txt");
+                    throw new FileExtensionException("Проверьте, что файл \"{0}\" имеет расширение {1} ", filename, FileExtensions.TXT);
+                    Log.Logger.Warning("Проверьте, что файл \"{0}\"  имеет расширение {1}",filename, FileExtensions.TXT);
                 }
             }
         }
-        public void Create(string filename, string content = null)
+        public void Create(string filename,string path, string content = null)
         {
-            string path = UserDirectory.Get();
+            if (string.IsNullOrWhiteSpace(path)) path = UserDirectory.Get();
             bool isNull = string.IsNullOrWhiteSpace(filename);
             if (!isNull)
             {
@@ -98,11 +99,11 @@ namespace EvidentInstruction.Models
                             try
                             {
                                 System.IO.File.Create(fullPath);
-                                Log.Logger.Information("Пустой файл был создан");
+                                Log.Logger.Information("Пустой файл \"{0}\" был создан в директории \"{1}\" ",filename, path);
                             }
                             catch (FileNotFoundException e)
                             {
-                                Log.Logger.Warning(e.Message);
+                                Log.Logger.Warning("Пустой файл \"{0}\" не был создан с ошибкой \"{1}\"", filename,e.Message);
                             }
                         }
                         else
@@ -110,11 +111,11 @@ namespace EvidentInstruction.Models
                             try
                             {
                                 System.IO.File.AppendAllText(fullPath, content);
-                                Log.Logger.Information("Файл был создан");
+                                Log.Logger.Information("Файл \"{0}\" был создан в директории \"{1}\" ", filename, path);
                             }
                             catch (FileNotFoundException e)
                             {
-                                Log.Logger.Warning(e.Message);
+                                Log.Logger.Warning("Пустой файл \"{0}\" не был создан с ошибкой \"{1}\"", filename, e.Message);
                             }
                         }
                     }
@@ -123,18 +124,18 @@ namespace EvidentInstruction.Models
                         try
                         {
                             System.IO.File.WriteAllText(fullPath, content);
-                            Log.Logger.Information("Файл был перезаписан");
+                            Log.Logger.Information("Файл \"{0}\" в директории \"{1}\" был перезаписан ", filename, path);
                         }
                         catch (FileNotFoundException e)
                         {
-                            Log.Logger.Warning(e.Message);
+                            Log.Logger.Warning("Файл \"{0}\" не был перезаписан с ошибкой \"{1}\"", filename, e.Message);
                         }
                     }
                 }
                 else
                 {
-                    throw new FileExtensionException("Файл " + filename + " не является текстовым файлом");
-                    Log.Logger.Warning("Файл " + filename + " не является текстовым файлом");
+                    throw new FileExtensionException("Файл \"{0}\" не является текстовым файлом", filename);
+                    Log.Logger.Warning("Файл \"{0}\" не является текстовым файлом", filename);
                 }
             }
             else
@@ -145,30 +146,34 @@ namespace EvidentInstruction.Models
         }
         public void Delete(string filename, string path)
         {
+            if (string.IsNullOrWhiteSpace(path)) path = UserDirectory.Get();
             string fullpath = Path.Combine(path, filename);
             string notfullpath = fullpath.Replace("\\", "");
-
+            if (string.IsNullOrWhiteSpace(filename))
+            {
+                Log.Logger.Warning("Имя файла отсутствует");
+                throw new NoFileNameException("Имя файла отсутствует");
+            }
             if (string.IsNullOrWhiteSpace(fullpath))
             {
-                Log.Logger.Warning("Путь до файла не указан");
-                throw new ArgumentNullException("Путь до файла " + path + " введен не верно");
+                Log.Logger.Warning("Путь до файла \"{0}\"  введен не верно", filename);
+                throw new ArgumentNullException("Путь до файла \"{0}\"  введен не верно", filename);
             }
             if (System.IO.File.Exists(fullpath)) System.IO.File.Delete(fullpath);
             else
             {
-                Log.Logger.Warning("Файла по указанному пути не существует");
-                throw new FileNotFoundException("Файла по указанному пути не существует");
+                Log.Logger.Warning("Файла \"{0}\" по указанному пути \"{1}\"  не существует",filename, path);
+                throw new FileExistException("Файла \"{0}\" по указанному пути \"{1}\"  не существует", filename, path);
             }
 
             if (IsExist(filename, path))
             {
-                Log.Logger.Warning("Файл " + filename + "не был удален");
-                throw new FileExistException("Файл " + filename + " не был удален");
+                Log.Logger.Warning("Файл \"{0}\" по указанному пути \"{1}\"  не  был удален", filename, path);
+                throw new FileExistException("Файл \"{0}\" по указанному пути \"{1}\"  не  был удален", filename, path);
             }
             else
             {
-                Log.Logger.Information("Файл " + filename + " был удален");
-                Console.WriteLine("Файл " + filename + "  был удален");
+                Log.Logger.Information("Файл \"{0}\" по указанному пути \"{1}\"   был успешно удален", filename, path);
             }
         }
         public string Get()
